feat: replaced HeaderProvider with layout duplication (not sure if its better)
