+++ src/app/[locale]/(main)/(modules)/view/[viewId]/page.tsx
