--- src/app/[locale]/(main)/page.tsx
@@ -1,12 +1,0 @@
-'use client';
-
-import { useEffect } from 'react';
-import { useShowSearchbar } from '@/components/header/HeaderProvider';
-
-export default function Home() {
-  const { setShowSearchbar } = useShowSearchbar();
-  useEffect(() => {
-    setShowSearchbar(false);
-  });
-  return <></>;
-}
